Move most globals back to sbws/globals.py
